fix the size_t problem.

+++ libEM/emdata_metadata.cpp
@@ -851,7 +851,7 @@ void EMData::set_size(int x, int y, int z)
 
 	int old_nx = nx;
 
-	size_t size = x*y*z*sizeof(float);
+	size_t size = (size_t)x*y*z*sizeof(float);
 
 	if (rdata != 0) {
 		rdata = (float*)EMUtil::em_realloc(rdata,size);
